requirements.txt
pandas==2.2.2
PyYAML==6.0.1
wheel==0.43.0
setuptools==69.5.1